Split library into cudf-free core and cucascade-cudf target
Decouple the core library from libcudf: cucascade now depends only on
RMM + CUDA (plus numa and kvikIO/cuFile for the disk tier), while the
cudf-backed representations, built-in converters, and bandwidth profiler
move to a separate cucascade-cudf target (headers under cucascade/cudf/).

- Generic memory::column_metadata (opaque int32_t type tag) used by the
  disk tier; cudf converters translate cudf::type_id <-> the tag.
- Virtualize record_writer_event/get_writer_event/rebind_stream on
  idata_representation so data_batch stays cudf-free.
- Core converter registry ships empty; register_builtin_converters() now
  lives in cucascade-cudf.
- CUCASCADE_BUILD_CUDF option (default ON) gates find_package(cudf) and
  the cudf target; cudf is exposed as a package component so core
  consumers never pull cudf.
- Split tests and benchmarks by target.
