+++ benchmark/benchmark_disk_converter.cpp
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-#include "cucascade/data/cpu_data_representation.hpp"
+#include "cucascade/cudf/builtin_converters.hpp"
+#include "cucascade/cudf/gpu_data_representation.hpp"
+#include "cucascade/cudf/host_data_representation.hpp"
 #include "cucascade/data/disk_data_representation.hpp"
-#include "cucascade/data/gpu_data_representation.hpp"
 #include "cucascade/data/representation_converter.hpp"
 #include "cucascade/memory/config.hpp"
 #include "cucascade/memory/memory_reservation_manager.hpp"
