+++ benchmark/benchmark_representation_converter.cpp
@@ -15,8 +15,9 @@
  * limitations under the License.
  */
 
-#include "cucascade/data/cpu_data_representation.hpp"
-#include "cucascade/data/gpu_data_representation.hpp"
+#include "cucascade/cudf/builtin_converters.hpp"
+#include "cucascade/cudf/gpu_data_representation.hpp"
+#include "cucascade/cudf/host_data_representation.hpp"
 #include "cucascade/data/representation_converter.hpp"
 #include "cucascade/memory/memory_reservation_manager.hpp"
