+++ cmake/cuCascadeConfig.cmake.in
@@ -2,11 +2,10 @@
 
 include(CMakeFindDependencyMacro)
 
-# Find required dependencies
+# Find required dependencies for the cudf-free core (RMM + CUDA + numa).
 find_dependency(CUDAToolkit REQUIRED VERSION 12.9)
 find_dependency(Threads REQUIRED)
 find_dependency(rmm REQUIRED CONFIG)
-find_dependency(cudf REQUIRED CONFIG)
 
 # The exported targets link the Numa::Numa imported target. Rerun the bundled
 # FindNuma module so it is reconstructed against the consumer's libnuma instead
@@ -14,11 +13,23 @@ find_dependency(cudf REQUIRED CONFIG)
 list(PREPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}/Modules")
 find_dependency(Numa REQUIRED)
 
+# cudf is only a dependency of the optional cucascade-cudf library, so it is
+# resolved lazily via the `cudf` package component. Plain
+# find_package(cuCascade) loads the cudf-free core and never pulls cudf;
+# find_package(cuCascade COMPONENTS cudf) additionally resolves cudf so the
+# cuCascade::cucascade_cudf target can be linked. cuCascade_cudf_FOUND reflects
+# whether the cudf-coupled library is available in this install.
+set(cuCascade_cudf_FOUND @CUCASCADE_BUILD_CUDF@)
+if(cuCascade_cudf_FOUND AND "cudf" IN_LIST cuCascade_FIND_COMPONENTS)
+  find_dependency(cudf REQUIRED CONFIG)
+endif()
+
 # Include the targets file
 include("${CMAKE_CURRENT_LIST_DIR}/cuCascadeTargets.cmake")
 
-# Provide a default target alias for convenience
-# Users can link to cuCascade::cucascade, cuCascade::cucascade_static, or cuCascade::cucascade_shared
+# Provide a default target alias for convenience Users can link to
+# cuCascade::cucascade, cuCascade::cucascade_static, or
+# cuCascade::cucascade_shared
 if(TARGET cuCascade::cucascade_shared)
   if(NOT TARGET cuCascade::cucascade)
     add_library(cuCascade::cucascade ALIAS cuCascade::cucascade_shared)
@@ -28,14 +39,31 @@ elseif(TARGET cuCascade::cucascade_static)
     add_library(cuCascade::cucascade ALIAS cuCascade::cucascade_static)
   endif()
 endif()
-# Users can link to cuCascade::cucascade_topology_discovery, cuCascade::cucascade_topology_discovery_static, or cuCascade::cucascade_topology_discovery_shared
+# Default alias for the cudf-coupled library (only present when
+# CUCASCADE_BUILD_CUDF was ON)
+if(TARGET cuCascade::cucascade_cudf_shared)
+  if(NOT TARGET cuCascade::cucascade_cudf)
+    add_library(cuCascade::cucascade_cudf ALIAS
+                cuCascade::cucascade_cudf_shared)
+  endif()
+elseif(TARGET cuCascade::cucascade_cudf_static)
+  if(NOT TARGET cuCascade::cucascade_cudf)
+    add_library(cuCascade::cucascade_cudf ALIAS
+                cuCascade::cucascade_cudf_static)
+  endif()
+endif()
+# Users can link to cuCascade::cucascade_topology_discovery,
+# cuCascade::cucascade_topology_discovery_static, or
+# cuCascade::cucascade_topology_discovery_shared
 if(TARGET cuCascade::cucascade_topology_discovery_shared)
   if(NOT TARGET cuCascade::cucascade_topology_discovery)
-    add_library(cuCascade::cucascade_topology_discovery ALIAS cuCascade::cucascade_topology_discovery_shared)
+    add_library(cuCascade::cucascade_topology_discovery ALIAS
+                cuCascade::cucascade_topology_discovery_shared)
   endif()
 elseif(TARGET cuCascade::cucascade_topology_discovery_static)
   if(NOT TARGET cuCascade::cucascade_topology_discovery)
-    add_library(cuCascade::cucascade_topology_discovery ALIAS cuCascade::cucascade_topology_discovery_static)
+    add_library(cuCascade::cucascade_topology_discovery ALIAS
+                cuCascade::cucascade_topology_discovery_static)
   endif()
 endif()
