+++ src/cudf/CMakeLists.txt
@@ -16,11 +16,8 @@
 # =============================================================================
 
 target_sources(
-  cucascade_objects
-  PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/disk_data_representation.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/io_backend_registry.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/pipeline_io_backend.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/data_batch.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/data_repository.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/data_repository_manager.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/representation_converter.cpp)
+  cucascade_cudf_objects
+  PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/gpu_data_representation.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/host_data_representation.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/bandwidth_profiler.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/representation_converter_builtins.cpp)
