+++ src/cudf/bandwidth_profiler.cpp
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-#include <cucascade/data/bandwidth_profiler.hpp>
-#include <cucascade/data/cpu_data_representation.hpp>
+#include <cucascade/cudf/bandwidth_profiler.hpp>
+#include <cucascade/cudf/gpu_data_representation.hpp>
+#include <cucascade/cudf/host_data_representation.hpp>
 #include <cucascade/data/disk_data_representation.hpp>
-#include <cucascade/data/gpu_data_representation.hpp>
 #include <cucascade/memory/chunked_resource_info.hpp>
 #include <cucascade/memory/memory_space.hpp>
