+++ src/cudf/host_data_representation.cpp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-#include <cucascade/data/cpu_data_representation.hpp>
+#include <cucascade/cudf/host_data_representation.hpp>
 
 #include <algorithm>
 #include <cstring>
