+++ test/utils/mock_test_utils.hpp
@@ -29,10 +29,6 @@
 #include <cucascade/memory/numa_region_pinned_host_allocator.hpp>
 #include <cucascade/memory/reservation_manager_configurator.hpp>
 
-#include <cudf/column/column_factories.hpp>
-#include <cudf/table/table.hpp>
-#include <cudf/types.hpp>
-
 #include <rmm/cuda_stream_view.hpp>
 #include <rmm/mr/cuda_memory_resource.hpp>
 #include <rmm/mr/per_device_resource.hpp>
@@ -144,65 +140,6 @@ inline std::vector<memory::memory_space_config> create_conversion_test_configs()
     .use_host_per_gpu()
     .set_per_host_capacity(4096ull * 1024 * 1024);
   return builder.build();
-}
-
-/**
- * @brief Create a simple cuDF table for testing.
- *
- * @param num_rows Number of rows in the table
- * @param num_columns Number of columns (1 or 2 supported)
- * @return cudf::table A simple table with numeric columns
- *
- * When num_columns == 1: Creates a single INT32 column filled with 0x42
- * When num_columns == 2: Creates INT32 (0x11) and INT64 (0x22) columns
- */
-inline cudf::table create_simple_cudf_table(
-  int num_rows,
-  int num_columns,
-  rmm::device_async_resource_ref mr = rmm::mr::get_current_device_resource_ref(),
-  rmm::cuda_stream_view stream      = rmm::cuda_stream_default)
-{
-  std::vector<std::unique_ptr<cudf::column>> columns;
-
-  // First column: INT32
-  auto col1 = cudf::make_numeric_column(
-    cudf::data_type{cudf::type_id::INT32}, num_rows, cudf::mask_state::UNALLOCATED, stream, mr);
-  if (num_rows > 0) {
-    auto view  = col1->mutable_view();
-    auto bytes = static_cast<size_t>(num_rows) * sizeof(int32_t);
-    CUCASCADE_CUDA_TRY(
-      cudaMemset(const_cast<void*>(view.head()), (num_columns == 1) ? 0x42 : 0x11, bytes));
-  }
-  columns.push_back(std::move(col1));
-
-  // Second column: INT64 (only if num_columns >= 2)
-  if (num_columns >= 2) {
-    auto col2 = cudf::make_numeric_column(
-      cudf::data_type{cudf::type_id::INT64}, num_rows, cudf::mask_state::UNALLOCATED, stream, mr);
-    if (num_rows > 0) {
-      auto view  = col2->mutable_view();
-      auto bytes = static_cast<size_t>(num_rows) * sizeof(int64_t);
-      CUCASCADE_CUDA_TRY(cudaMemset(const_cast<void*>(view.head()), 0x22, bytes));
-    }
-    columns.push_back(std::move(col2));
-  }
-
-  return cudf::table(std::move(columns));
-}
-
-inline cudf::table create_simple_cudf_table(
-  int num_rows,
-  rmm::device_async_resource_ref mr = rmm::mr::get_current_device_resource_ref(),
-  rmm::cuda_stream_view stream      = rmm::cuda_stream_default)
-{
-  return create_simple_cudf_table(num_rows, 2, mr, stream);
-}
-
-inline cudf::table create_simple_cudf_table(
-  rmm::device_async_resource_ref mr = rmm::mr::get_current_device_resource_ref(),
-  rmm::cuda_stream_view stream      = rmm::cuda_stream_default)
-{
-  return create_simple_cudf_table(100, 2, mr, stream);
 }
 
 }  // namespace test
